Consolidate token form models into one file
